Move ReaderCardUiState into its own class
