+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/ReaderDiscoverViewModel.kt
@@ -1,8 +1,6 @@
 package org.wordpress.android.ui.reader.discover
 
-import android.text.Spanned
 import android.view.View
-import androidx.annotation.AttrRes
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import kotlinx.coroutines.CoroutineDispatcher
@@ -12,13 +10,9 @@ import org.wordpress.android.modules.UI_THREAD
 import org.wordpress.android.ui.reader.ReaderTypes.ReaderPostListType.TAG_FOLLOWED
 import org.wordpress.android.ui.reader.discover.ReaderDiscoverViewModel.DiscoverUiState.ContentUiState
 import org.wordpress.android.ui.reader.discover.ReaderDiscoverViewModel.DiscoverUiState.LoadingUiState
-import org.wordpress.android.ui.reader.models.ReaderImageList
 import org.wordpress.android.ui.reader.repository.ReaderPostRepository
-import org.wordpress.android.ui.utils.UiDimen
-import org.wordpress.android.ui.utils.UiString
 import org.wordpress.android.util.AppLog
 import org.wordpress.android.util.AppLog.T
-import org.wordpress.android.util.image.ImageType
 import org.wordpress.android.viewmodel.ScopedViewModel
 import javax.inject.Inject
 import javax.inject.Named
@@ -132,64 +126,5 @@ class ReaderDiscoverViewModel @Inject constructor(
         data class ContentUiState(val cards: List<ReaderCardUiState>) : DiscoverUiState(contentVisiblity = true)
         object LoadingUiState : DiscoverUiState(progressVisibility = true)
         object ErrorUiState : DiscoverUiState()
-    }
-
-    sealed class ReaderCardUiState {
-        data class ReaderPostUiState(
-            val postId: Long,
-            val blogId: Long,
-            val dateLine: String,
-            val title: String?,
-            val blogName: String?,
-            val excerpt: String?, // mTxtText
-            val blogUrl: String?,
-            val photoTitle: String?,
-            val featuredImageUrl: String?,
-            val featuredImageCornerRadius: UiDimen,
-            val fullVideoUrl: String?,
-            val avatarOrBlavatarUrl: String?,
-            val thumbnailStripSection: GalleryThumbnailStripData?,
-            val discoverSection: DiscoverLayoutUiState?,
-            val videoOverlayVisibility: Boolean,
-            val moreMenuVisibility: Boolean,
-            val photoFrameVisibility: Boolean,
-            val bookmarkAction: ActionUiState,
-            val likeAction: ActionUiState,
-            val reblogAction: ActionUiState,
-            val commentsAction: ActionUiState,
-            val postHeaderClickData: PostHeaderClickData?,
-            val onItemClicked: (Long, Long) -> Unit,
-            val onItemRendered: (Long, Long) -> Unit,
-            val onMoreButtonClicked: (Long, Long, View) -> Unit,
-            val onVideoOverlayClicked: (Long, Long) -> Unit
-        ) : ReaderCardUiState() {
-            val dotSeparatorVisibility: Boolean = blogUrl != null
-
-            data class PostHeaderClickData(
-                val onPostHeaderViewClicked: ((Long, Long) -> Unit)?,
-                @AttrRes val background: Int
-            )
-
-            data class GalleryThumbnailStripData(
-                val images: ReaderImageList,
-                val isPrivate: Boolean,
-                val content: String // needs to be here as it's required by ReaderThumbnailStrip
-            )
-
-            data class DiscoverLayoutUiState(
-                val discoverText: Spanned,
-                val discoverAvatarUrl: String,
-                val imageType: ImageType,
-                val onDiscoverClicked: ((Long, Long) -> Unit)
-            )
-
-            data class ActionUiState(
-                val isEnabled: Boolean,
-                val isSelected: Boolean = false,
-                val contentDescription: UiString? = null,
-                val count: Int = 0,
-                val onClicked: ((Long, Long, Boolean) -> Unit)? = null
-            )
-        }
     }
 }
